update: refactoring structure of the project and implement new component text
