+++ src/components/parts/label/label.scss
