+++ src/components/parts/text/text.scss
@@ -1,0 +1,11 @@
+.text{
+    padding-left: 5px;
+    margin-top: 10px;
+    font-size: 1rem;
+    font-family: 'Roboto';
+    font-weight: 500;
+}
+
+.text-danger{
+    color: #bd2130;
+}
